refactor: extract useTransactionActions from Dashboard
Extract transaction operations into useTransactionActions hook (1043 lines).
Removes ~945 lines from Dashboard.js (2889 -> 2010). Largest extraction.

Hook contains: loadMoreHistoricalTransactions, loadPastTransactions,
loadMoreMonths, getDateRangePresets, parseCreatedAt, parseTransactionDate,
filterTransactionsByDateRange, loadTransactionsForDateRange,
handleCustomDateRange, handleClearDateFilter, getFilteredStats,
filterTransactionsBySearch, getDisplayTransactions, handleViewTransition,
handleRefresh, exportFullTransactions, exportBasicTransactions,
convertTransactionsToBasicCSV, downloadCSV, triggerDownload,
handleTxSearch* handlers, groupTransactionsByMonth, getMonthGroups,
toggleMonth, loadMoreTransactionsForMonth.

Also removes now-unused getEnvironment and getApiUrl imports from Dashboard.
